Dump fragment lengths to sidecar file
This is part of a solution to recover original fragment spans.

In some situations, a stream may be rotated while maintaining the
total number of pixels (e.g. 1920x1080 to 1080x1920 for shorts).

YouTube does not issue a new manifest in this situation, and the
format itag is unchanged, so this tool continues to append
fragments to an existing file.

ffmpeg appears to take the initially found dimensions and applies
them to all subsequent fragments on *any* output operation, and
with h264 this causes fragments with 'fixed' dimensions to be
improperly decoded, even if you do a fast seek on the TS file.

Because we append downloaded fragments as-is, we can build a tool
to pull untouched fragment spans post-download.

Without diving into decoding the TS segment in memory to decide
which file to append to, one alternative is to dump fragments
into individual files and concat them during the final mux
instead.  One positive is that fragment downloads could be
idempotent and would not need to be synchronized for appending to
One Large File, though one negative is not being able to easily
invoke ffmpeg on a 'live' file for spot checking.

I think we'll just opt to take this route for now.  Feels simpler.
We may decode the segment in memory using something like PyAV
later.