Prevent ffmpeg from treating files as flags
ffmpeg does not like that we set the YouTube video ID as the
basename of the file, thinking that it's supposed to act as an
option due to the leading hyphen.

We work around this by resolving to an absolute path so the
path arguments always start with '/' or a drive letter.