Improve ffmpeg perf by suppressing VP9 error flood
Remux with VP9 video streams causes the following error:

> Found unknown-length element with ID 0x18538067 at pos. 0x15fa07a9e
> for which no syntax for parsing is available.

This does apply to ffmpeg 7.0, possibly older versions.

There doesn't seem to be any good way to squash this error in
particular without getting into the machinery of parsing Matroska
containers in general, so we just raise the log level again for
now.

This is also what ytarchive does, so we'll settle for that.