Allow multiple itag-fragment downloads in flight
By making multiple requests at once for fragments that are known
to exist (behind the live seqnum), we get a modest increase in
thoroughput so long as we aren't being throttled.

Most of the control flow is unchanged - the fragments themselves
are still processed serially when being written to the output file,
and the rest of the tasks are discarded if the current one fails
for some reason.

That latter bit is useful in our case since a failure indicates a
good chance that we need to refetch an expired manifest.