Add xerror flag to ffmpeg
-xerror stops the ffmpeg process and exits on errors.

It turns out that ffmpeg's default configuration may silently
accept major system issues like I/O errors (which I happened to
discover while running a test instance of moombox on a failing
hard disk).

We also add a bitstream filter to ensure that the DTS is
monotonically increasing to process the error case before ffmpeg
does (and with -xerror, explodes).

Closes #11.